+++ packages/fluentui/react-bindings/src/hooks/useStyles.ts
@@ -36,7 +36,8 @@ const defaultContext: StylesContextValue<{ renderRule: RendererRenderRule }> = {
   performance: {
     enableSanitizeCssPlugin: process.env.NODE_ENV !== 'production',
     enableStylesCaching: true,
-    enableVariablesCaching: true
+    enableVariablesCaching: true,
+    enableHardVariablesCaching: false
   },
   renderer: { renderRule: () => '' },
   theme: emptyTheme