+++ packages/fluentui/react-bindings/src/styles/types.ts
@@ -66,6 +66,7 @@ export interface StylesContextPerformance {
   enableSanitizeCssPlugin: boolean;
   enableStylesCaching: boolean;
   enableVariablesCaching: boolean;
+  enableHardVariablesCaching: boolean;
 }
 
 export type StylesContextPerformanceInput = Partial<StylesContextPerformance>;